boot: Remove old launch junk
Long ago, most of pox.py was moved to boot.py.  The launch junk (for
picking the interpreter, etc.) went with it, even though I don't
think boot.py is actually runnable on its own (there's a comment
about fixing that, but it hasn't happened).  And there's no known
use case for running it with "./", so we'll just drop the launch
junk from here.